Parsing fence files
This is not an efficient implementation, and relies on a mixture of serde gencode and handwritten
deserialisers. Unfortunately the good-fences fence specification is incredibly permissive,
requiring the deserialising party to react to changes in the structure at parse-time.

Serde can't handle that for you with gencode, so I had to jump through a bunch of hoops to get
this to work.

It's my first time doing anything complicated in serde so I imagine it is very wrong.