Matching ignored dirs will be added as source files with empty impot map (#26)
Currently, `ignoredDirs` option only excludes files from being evaluated
with their fence config. The problem is that even when files in
`shared/.**/lib` or `dist` are not evaluated, their AST is still being
parsed, considering that it is an expensive task to do, it is having an
impact on performance.

- Files under matching `ignoredDirs` regexes will be ignored (sources
and fences).
- Resolved import paths matching `ignoredDirs` are ignored, even if they
would be violations.