Restructure logger to enable `logger_srcfile` (#104)
The overall goal of this change is to support a "sub-trait" of `logger`
that adds some src-file logging methods associated against a given
logging.

This will be used to trace issues with src files while e.g. parsing or
interpreting the structure of source files. As part of this change, the
`Sync` requirement of logger had to be dropped.
This is because the src file informatio nwe get from `swc` is
necessarially non-`Sync`/`Send`.

---------

Co-authored-by: Max Huang-Hobbs <Maxwell.HuangHobbs@microsoft.com>