devcontainer.json: use slirp4netns to boot offline
replaces the default (paste) with slirp4netns in the podman container.

This is a workaround to allow using the container when
offline.

Note that this actually degrades network peformance for the container,
so it should be removed once the fix for pasta/podman is eventually
implemented.
See: https://github.com/containers/podman/issues/23566

Note that this requires the host system to have slirp4netns installed
on fedora you can do this via `dnf install slirp4netns`