Fix infinite loop of expansion during resolution
In f66e808e16bd713ac48a7579b38286e5ffd16fbc, I added a frontier walk to
expand recursive star imports would be resolved to a concrete set of
names.

However, this had a bug where if a star import into an external package
was visited, it would avoid expanding it, and infinitely loop after
seeing remaining "*" exports to expand.

This commit fixes that by ensuring that we only loop when the expanded
exports contain a '*', not when the file contains an unresolved '*'
export